requirements.txt
discord.py-self
pymongo
mysql-connector-python
redis
